- Move read_stdin() / MAX_STDIN_BYTES from vitest.rs and pytest.rs into shared.rs as read_stdin_raw(); remove the four duplicated inline copies - Update vitest.rs (passthrough + main paths) and pytest.rs (passthrough + main paths) to use try_read_stdin; pytest wraps the raw String into a synthetic CommandOutput inline — same behaviour, zero duplicate logic - Replace the inlined !stdin().is_terminal() && args.is_empty() in cargo.rs with shared::should_read_stdin(args); remove the now-unused IsTerminal import - Move the 3 test_should_read_stdin_* unit tests from vitest.rs to shared.rs::tests where the function is defined Co-Authored-By: Claude <noreply@anthropic.com> EOF )
